move lsm to leveldb
