+++ src/leveldb/error.rs
