+++ src/leveldb/util.rs
